Major changes to game engine (#55)

+++ game/graphics/internal/uniform_material.go
@@ -10,6 +10,6 @@ func (u MaterialUniform) Std140Plot(plotter *blob.Plotter) {
 	plotter.PlotBytes(u.Data)
 }
 
-func (u MaterialUniform) Std140Size() int {
-	return len(u.Data)
+func (u MaterialUniform) Std140Size() uint32 {
+	return uint32(len(u.Data))
 }
